fix: update on project card
