+++ src/components/Projects/ProjectCard/ProjectCard.css
@@ -45,15 +45,6 @@
   width: 100%;
 }
 
-.ProjectCard .ProposalUpdate {
-  border: none;
-  border-top: 1px solid var(--black-100);
-  border-radius: 0px;
-}
-.ProjectCard .ProposalUpdate:last-child {
-  margin-bottom: 0;
-}
-
 @media (min-width: 425px) {
   .ProjectCard {
     padding: 0 !important;
